clearer separation of concerns when bootstrapping application
